Rename ND_FUNC to ND_CALL

+++ 9cclib.c
@@ -9,11 +9,10 @@ int ret33() {
     return 33;
 }
 
-int multi2(long a) {
+int multi2(int a) {
     return a * 2;
 }
 
-long sum(long a, long b, long c, long d, long e, long f) {
+int sum(int a, int b, int c, int d, int e, int f) {
     return a + b + c + d + e + f;
 }
-
